+++ CHANGELOG.md
@@ -2,3 +2,8 @@
 
 # v1.0.0
 - Initial Release
+
+# v1.3.0
+- Add Server Name Indication (SNI) support
+- Add tests to cover each root certificate authority in use
+- Fix MITM proxy error reporting